footer: Remove copyright year
The current copyright year appearing in the footer (2021) is outdated.
We could replace it, but as Liz observed, the CNCF recommends dropping
it [0]. Let's update the footer to remove the copyright year. I think
that the "All rights reserved." may be covered by the Copyright mention
and could maybe be removed as well, but in doubt, I didn't touch that
part.

[0] https://github.com/cncf/foundation/blob/main/copyright-notices.md#copyright-notices